changes to make tests pass on 1.8.7
The main changes were that in 1.9+ you can chain commands with the dot
starting a new line, but you can't in 1.8.  So this:

    [1, 2, 3, 4].select { |x| x < 3 }
                .map { |x| x * x }

Is not acceptable to 1.8.  The other main change is that JSON wasn't
recognized, so I had to add it as an explicit dependency.  I also had
to add rake as a development dependency, thought that could just be me.
And then one array was out of order in a test, but that made for a
better test anyway.